requirements.txt
Flask==2.0.1
openai==0.28.0
python-dotenv==0.19.0
gunicorn==20.1.0